Information_Project/requirements.txt
django
imdbpy
django-crispy-forms
django-allauth
sklearn
pandas
